Composability: let the NFT asset react to permissions being set
If the NFT has state that depends on the SACD permissions, currently there's no way for it to know the permissions changed (before someone calling it after permission change). This is similar to a contract receiving ERC20 tokens: the contract doesn't get notified. It's in contrast to a contract receiving ether: the receiving contract gets a chance to update its state (fallback / receive function). This is quite beneficial to composing smart contracts: if setting SACD permission activates the target NFT as well, it can keep its own state updated, so that all contracts stay in sync.